test(web-fetch): colocate extractor coverage with runtime owner
